Copy and paste all from the catch-the-bananas branch, excluding the tests
All of this comes from the catch-the-bananas branch. There were some unknown, unexpected errors occurring in that branch, despite it compiling without errors just before. It is presumably caused by the assembly definition file while creating tests. I have decided to copy and paste the files handling the functionality of Catch The Bananas to start over.
To view the testing for this feature, you may go to the catch-the-bananas branch and go to Assets\CTBTests. I decided I will not bring the testing into this branch as it causes errors I cannot solve as of now.